backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
bcrypt==4.1.1
PyJWT==2.8.0
Flask-JWT-Extended==4.5.2
